@@ -19,9 +19,7 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>11</java.version>
-
-		<spring.profiles.active>dev</spring.profiles.active>
+		<java.version>17</java.version>
 	</properties>
 
 	<dependencies>
