@@ -1,7 +1,7 @@
 package de.promove.autokss.dao;
 
-import javax.persistence.criteria.JoinType;
-import javax.persistence.metamodel.Attribute;
+import jakarta.persistence.criteria.JoinType;
+import jakarta.persistence.metamodel.Attribute;
 import java.util.ArrayList;
 import java.util.List;
 
