@@ -3,17 +3,12 @@ package de.promove.autokss.web.crud;
 import de.promove.autokss.model.User;
 import de.promove.autokss.service.UserService;
 import de.promove.autokss.web.common.crud.AbstractCrudBean;
-import de.promove.autokss.web.util.GrowlMessenger;
-import de.promove.autokss.web.util.MessageFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Scope;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Controller;
 
-import javax.el.MethodExpression;
-import javax.faces.bean.ManagedBean;
-import javax.faces.view.ViewScoped;
+import jakarta.faces.bean.ManagedBean;
+import jakarta.faces.view.ViewScoped;
 
 @Component
 @ManagedBean
