@@ -5,8 +5,8 @@ import org.springframework.security.core.userdetails.UserCache;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.stereotype.Service;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 //@Service
 public class CustomLogoutHandler implements LogoutHandler {
