@@ -24,7 +24,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         // require all requests to be authenticated except for the resources
-        http.authorizeRequests().antMatchers("/javax.faces.resource/**", "/*")
+        http.authorizeRequests().antMatchers("/jakarta.faces.resource/**", "/*")
                 .permitAll().anyRequest().authenticated();
         // login
         http.formLogin().loginPage("/login.xhtml").permitAll().defaultSuccessUrl("/index.xhtml")
